pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skaro"
version = "0.1.9"
description = "AI-powered SDLC orchestration platform"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
authors = [
    { name = "Skaro Contributors" },
]
keywords = ["ai", "development", "llm", "specification", "methodology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "questionary>=2.0",
    "pyyaml>=6.0",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "websockets>=13.0",
    "watchfiles>=0.24",
    "gitpython>=3.1",
    "jinja2>=3.1",
    "mistune>=3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]

[project.scripts]
skaro = "skaro_cli.main:cli"

[project.urls]
Homepage = "https://github.com/skarodev/skaro"
Repository = "https://github.com/skarodev/skaro"

[tool.hatch.build.targets.wheel]
packages = ["src/skaro_core", "src/skaro_cli", "src/skaro_web"]

# Bundle root-level templates/ into the skaro_core package so that
# importlib.resources.files("skaro_core") / "templates" resolves correctly
# after pip install.
[tool.hatch.build.targets.wheel.force-include]
"templates" = "skaro_core/templates"

[tool.hatch.build.hooks.custom]
path = "build_hooks/frontend_hook.py"

[tool.hatch.build.targets.sdist]
include = [
    "src/",
    "templates/",
    "frontend/",
    "build_hooks/",
    "tests/",
    "docs/",
    "pyproject.toml",
    "README.md",
    "LICENSE",
    "MANIFEST.in",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"